refactor: rename files according to Nest.js architecture conventions
