+++ src/artist/errors/incorrect-artist-id.error.ts
