+++ src/artist/utils/uuid-artist-validate.util.ts
