+++ src/user/utils/prepare-user-for-response.util.ts
